requirements.txt
numpy==1.24.3
numpy==1.17.4
Pillow==10.1.0
Pillow==7.0.0
Pillow==10.2.0
streamlit==1.29.0
tensorflow==2.13.1
tensorflow_hub==0.15.0
